+++ CHANGELOG.md
@@ -5,9 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to 
 [Semantic Versioning](http://semver.org/).
 
-## 0.3.0.dev (development stage/unreleased/unstable)
+## 0.4.0.dev (development stage/unreleased/unstable)
+
+## 0.4.0
 ### CHANGED
-- The command line interface `ubtsl`
+- Many updates in the command line interface `ubtsl`
 
 ## 0.3.0
 ### Added